Splitted some code
