+++ new-client/src/plugins/Measurer/Measurer.js
@@ -1,4 +1,4 @@
-import React, { useRef } from "react";
+import React, { useRef, useMemo } from "react";
 import { useState, useCallback } from "react";
 import BaseWindowPlugin from "../BaseWindowPlugin";
 
@@ -11,10 +11,7 @@ import { DEFAULT_MEASUREMENT_SETTINGS } from "./constants";
 import DrawModel from "models/DrawModel";
 import { Circle, Fill, RegularShape, Stroke, Style } from "ol/style";
 import HelpIcon from "@mui/icons-material/Help";
-import { Feature } from "ol";
-import { LineString } from "ol/geom";
-import { lineString as TurfLineString } from "@turf/helpers";
-import booleanPointOnLine from "@turf/boolean-point-on-line";
+import AngleSnapping from "./AngleSnapping";
 
 function Measurer(props) {
   const { map, app } = props;
@@ -25,7 +22,9 @@ function Measurer(props) {
   const [pluginShown, setPluginShown] = React.useState(
     props.options.visibleAtStart ?? false
   );
-  const snapGuides = React.useRef([]);
+  const angleSnapping = useMemo(() => {
+    return new AngleSnapping();
+  }, []);
 
   const [drawModel] = React.useState(
     () =>
@@ -59,154 +58,12 @@ function Measurer(props) {
       })
   );
 
-  const clearSnapGuides = useCallback(() => {
-    snapGuides.current.forEach((guideFeature) => {
-      drawModel.removeFeature(guideFeature);
-    });
-    snapGuides.current = [];
-  }, [drawModel]);
-
   const handleDrawStart = useCallback(
     (e) => {
-      if (!e.feature) {
-        return;
-      }
-
-      const measureFeature = e.feature;
-
-      // The snapping angles
-      // Maybe we should be able to configure these in the future
-      const anglesToGenerate = [0, 90, 180, 270];
-
-      // We can only create perpendicular snapping for these types.
-      const allowedTypes = ["Polygon", "LineString"];
-      const featureType = measureFeature.getGeometry().getType();
-
-      if (!allowedTypes.includes(featureType)) {
-        // Lets escape
-        return;
-      }
-
-      // We'll need the clicked coordinate for lookup
-      const coord = drawModel.getFeatureCoordinates(measureFeature)[0];
-
-      if (coord) {
-        // Unfortunately we need to convert to pixel because there is no "getFeaturesAtCoordinate" at this point
-        const px = map.getPixelFromCoordinate(coord);
-        const clickedFeatures = map
-          .getFeaturesAtPixel(px, {
-            hitTolerance: 0,
-          })
-          .filter((f) => {
-            // Filter out our draw feature and circles, points etc.
-            const type = f.getGeometry().getType();
-            return f !== measureFeature && allowedTypes.includes(type);
-          });
-
-        if (clickedFeatures.length === 0) {
-          // No useful features found
-          return;
-        }
-
-        // Lets use the first available feature and get all its coordinates
-        let allCoordinates = drawModel.getFeatureCoordinates(
-          clickedFeatures[0]
-        );
-
-        let i = 0;
-        let clickedSegment = null;
-
-        // If a geometry is too complex we might choke the browser....
-        // We'll prevent this by limiting this functionality to more simple geometries.
-        const maxSegmentsToCheck = 500; // 500 is still allot....
-
-        while (
-          !clickedSegment &&
-          i < allCoordinates.length &&
-          i <= maxSegmentsToCheck
-        ) {
-          const segment = [allCoordinates[i], allCoordinates[i + 1]];
-
-          // getClosestPoint did not work as expected so I used methods from turf
-          let foundClickedSegment = booleanPointOnLine(
-            coord,
-            new TurfLineString(segment),
-            {
-              ignoreEndVertices: false,
-              epsilon: 0.001, // I got no matches without a value here (hitTolerance/fuzziness)
-            }
-          );
-
-          if (foundClickedSegment) {
-            clickedSegment = segment;
-          }
-
-          i++;
-        }
-
-        if (clickedSegment) {
-          const segmentLine = new LineString(clickedSegment);
-
-          // Get the angle of the clicked segment
-          const angleRadians = Math.atan2(
-            clickedSegment[0][0] - clickedSegment[1][0],
-            clickedSegment[0][1] - clickedSegment[1][1]
-          );
-
-          // Lets add a nice line on top of the segment
-          // to highlight the "owner" segment. Kind of like a selection.
-          const segmentStroke = new Stroke({
-            color: "rgba(0, 255, 0, 0.2)",
-            width: 5,
-          });
-          const segmentFeature = new Feature({
-            geometry: segmentLine,
-          });
-          segmentFeature.setStyle(
-            new Style({
-              stroke: segmentStroke,
-            })
-          );
-
-          drawModel.addFeature(segmentFeature);
-
-          // We'll use this style for all added snapping lines.
-          const guideStyle = new Style({
-            stroke: new Stroke({
-              color: "rgba(0, 255, 0, 0.5)",
-              lineDash: null, // Warning! Adding lineDash here makes this sooooooo slow.
-              width: 1,
-            }),
-          });
-
-          let guides = [];
-
-          anglesToGenerate.forEach((angle) => {
-            angle = (angle * Math.PI) / 180; // Degrees to radians
-            let targetX = coord[0] + 100000 * Math.cos(angle);
-            let targetY = coord[1] + 100000 * Math.sin(angle);
-            let guideLine = new LineString([
-              [coord[0], coord[1]],
-              [targetX, targetY],
-            ]);
-            guideLine.rotate(-angleRadians, coord);
-            let feature = new Feature({
-              geometry: guideLine,
-            });
-            feature.setStyle(guideStyle);
-
-            guides.push(feature);
-            drawModel.addFeature(feature);
-          });
-
-          // Add the guides etc to ref arr so we can remove them
-          // later in clearSnapGuides() (drawEnd)
-          guides.push(segmentFeature);
-          snapGuides.current = guides;
-        }
-      }
+      // Forward drawstart event etc. to angle snapper.
+      angleSnapping.handleDrawStartEvent(e, map, drawModel);
     },
-    [map, drawModel, snapGuides]
+    [angleSnapping, map, drawModel]
   );
 
   const handleAddFeature = useCallback(
@@ -239,7 +96,7 @@ function Measurer(props) {
 
   const handleDrawEnd = useCallback(
     (e) => {
-      clearSnapGuides();
+      angleSnapping.clearSnapGuides(drawModel);
       if (!e.feature) return;
       const feature = e.feature;
       const type = feature.getGeometry().getType();
@@ -272,7 +129,7 @@ function Measurer(props) {
 
       feature.setStyle(style);
     },
-    [clearSnapGuides]
+    [angleSnapping, drawModel]
   );
 
   const startInteractionWithDrawType = useCallback(
